Changed code around but project still works the same
I tried changing the code around where you upload images, but I think the thing that makes my uploader not work is in index. When I try initializing $filename in a for loop, the uploader seems to sort of work but it loops everything a bunch of times. Seems I need to find a more stable middleground for this to work, or something.